Fix parsing the docs after removing docs from the local repository
Removing the docs from the api-docs folder meant that we can't parse 
them there. With this change, the documentation is parsed in the output 
folder, so it relies on the copy_files task being completed. This also 
switches to using a task instead of hacking set_site to perform the 
operation early.