Use the checksum option to compute file changes
It seems there are permissions and timestamp changes in the 
"documentation/docs-*" folders, presumably because they're copied into 
the output freshly on every build and the remote user is different than 
the local user. The checksum option computes the MD5 hash of each file 
to compare, reducing the number of files transferred in this case. The 
performance penalty of computing all these checksums seems to be minimal 
for our case.