Copy relevant bits of late_load_js macro
The main branch of Nikola has changed the URL for jQuery to https, but
it hasn't been released yet and we're pinned to 8.1.1 for the moment.
This change copies the relevant Javascript loading lines from the
updated template files in Nikola.

Also update the date-fanciness Javascript to match the 8.1.1 release.